determined that the slowdown is definitely due to the large numer of async functions being called and waited on in Promise.all. so, the thing to do is to not call an async function when we need to, and we've made a preliminary stab at it and made the methods that we'll need to check if a derived instance actually has implemented the event listener, but the main bulk of the problem is in how we call children's event listeners. it's a difficult problem that will require some thought to solve

+++ js/event.js
@@ -20,8 +20,9 @@ export class Grid {
         return this.objects.has(object);
     }
     Add(object) {
-        Utils.Assert(!this.Has(object), `This object is already in the grid.`);
-        this.objects.set(object, new Listeners());
+        if (!this.Has(object)) {
+            this.objects.set(object, new Listeners());
+        }
     }
     Add_Many(objects) {
         for (const object of objects) {
@@ -29,9 +30,10 @@ export class Grid {
         }
     }
     Remove(object) {
-        Utils.Assert(this.Has(object), `This object is not in the grid.`);
-        this.Remove_All_Listeners(object);
-        this.objects.delete(object);
+        if (this.Has(object)) {
+            this.Remove_All_Listeners(object);
+            this.objects.delete(object);
+        }
     }
     Remove_Many(objects) {
         for (const object of objects) {
