+++ ts/event.ts
@@ -28,12 +28,9 @@ export class Grid
     ):
         void
     {
-        Utils.Assert(
-            !this.Has(object),
-            `This object is already in the grid.`,
-        );
-
-        this.objects.set(object, new Listeners());
+        if (!this.Has(object)) {
+            this.objects.set(object, new Listeners());
+        }
     }
 
     Add_Many(
@@ -51,13 +48,10 @@ export class Grid
     ):
         void
     {
-        Utils.Assert(
-            this.Has(object),
-            `This object is not in the grid.`,
-        );
-
-        this.Remove_All_Listeners(object);
-        this.objects.delete(object);
+        if (this.Has(object)) {
+            this.Remove_All_Listeners(object);
+            this.objects.delete(object);
+        }
     }
 
     Remove_Many(
